relocate FileManager to utils
